requirements.txt
playwright==1.40.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.12.0
chromadb==0.4.22
sentence-transformers==2.2.2
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic>=2.0.0 